Data on policy area histories

This dataset includes data on the histories of the policy areas covered by the departments of the Commission (1958-2021). There is one observation per historical interval in which the mapping between a policy area and a department is constant. If the nomenclature of a department changes, a new interval will be created. Unlike allocations, intervals can span Commissions. There can also be multiple overlaping intervals per policy area when a policy area is split across multiple departments at the same time. By sorting the dataset by policy area, you can see the evolution of the mapping between the policy area and departments over time. The dataset indicates the Commissions associated with each interval and the start and end dates of each interval.

policy_area_histories

Format

A data frame with 18 variables:

key_id

Numeric. An ID number that uniquely identifies each observation. Indicates the default sort order for the dataset.

policy_area_id

Numeric. An ID number that uniquely identifies each policy area. See the policy_areas dataset.

policy_area

String. The name of the policy area. See the policy_areas dataset.

department_id

Numeric. An ID number that uniquely identifies each department. See the departments dataset.

department_name

String. The name of the department.

department_code

String. The common abbreviation of the department. Coded Not applicable of there is not an abbreviation for the department. Coded None for all Commissions prior to the Prodi Commission.

department_type_id

Numeric. An ID number that uniquely identifies the type of the department. Coded 1 for Directorate-General and 2 for Service department. See the variable department_type.

department_type

String. The type of the department. Either Directorate-General or Service department.

commissions

String. A list of the Commissions in which the policy area appeared, separated by a comma.

count_commissions

Numeric. The number of Commissions in which the policy area appeared.

start_date

Date. The start date of the historical interval. Historical intervals can cross Commissions.

start_year

Numeric. The start year of the historical interval. Historical intervals can cross Commissions.

start_month

Numeric. The start month of the historical interval. Historical intervals can cross Commissions.

start_day

Numeric. The start day of the historical interval. Historical intervals can cross Commissions.

end_date

Date. The end date of the historical interval. Historical intervals can cross Commissions.

end_year

Numeric. The start year of the historical interval. Historical intervals can cross Commissions.

end_month

Numeric. The start month of the historical interval. Historical intervals can cross Commissions.

end_day

Numeric. The end day of the historical interval. Historical intervals can cross Commissions.