Data on policy areas

This dataset includes data on policy areas covered by departments of the Commission (1958-2021). There is one observation per policy area. EUR-Lex, the Commission's official database of EU legal documents, sorts documents based on a case law directory and a directory of legal acts. There is a pre-Lisbon and post-Lisbon version of the case law directory. The dataset indicates the directory codes that correspond to each policy area according to each of these classification schemes.

policy_areas

Format

A data frame with 6 variables:

key_id

Numeric. An ID number that uniquely identifies each observation. Indicates the default sort order for the dataset.

policy_area_id

Numeric. An ID number that uniquely identifies each policy area. See the policy_areas dataset.

policy_area

String. The name of the policy area. See the policy_areas dataset.

case_law_old

String. A list of the codes from the pre-Lisbon version of EUR-Lex's Case Law Directory that apply to the policy area.

case_law_new

String. A list of the codes from the post-Lisbon version of EUR-Lex's Case Law Directory that apply to the policy area.

legislation

String. A list of the codes from EUR-Lex's Directory of Legal Acts that apply to the policy area.